requirements.txt
Flask==2.3.3
opencv-python-headless==4.10.0.84
numpy<2.0.0
matplotlib==3.7.2
Pillow==10.0.0
scikit-image==0.21.0